+++ crates/sui-indexer/src/store/mod.rs
@@ -2,19 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub(crate) use indexer_analytical_store::*;
-pub use indexer_store::*;
 pub(crate) use indexer_store_v2::*;
 pub use pg_indexer_analytical_store::PgIndexerAnalyticalStore;
-pub use pg_indexer_store::PgIndexerStore;
 pub use pg_indexer_store_v2::PgIndexerStoreV2;
 
 mod indexer_analytical_store;
-mod indexer_store;
 pub mod indexer_store_v2;
-pub mod module_resolver;
-pub(crate) mod module_resolver_v2;
+pub mod module_resolver_v2;
 mod pg_indexer_analytical_store;
-mod pg_indexer_store;
 mod pg_indexer_store_v2;
 mod pg_partition_manager;
 mod query;
@@ -28,18 +23,6 @@ pub(crate) mod diesel_macro {
                 .read_only()
                 .run($query)
                 .map_err(|e| IndexerError::PostgresReadError(e.to_string()))
-        }};
-    }
-
-    macro_rules! transactional_blocking {
-        ($pool:expr, $query:expr) => {{
-            let mut pg_pool_conn = crate::get_pg_pool_connection($pool)?;
-            pg_pool_conn
-                .build_transaction()
-                .serializable()
-                .read_write()
-                .run($query)
-                .map_err(|e| IndexerError::PostgresWriteError(e.to_string()))
         }};
     }
 
@@ -77,6 +60,5 @@ pub(crate) mod diesel_macro {
     }
 
     pub(crate) use read_only_blocking;
-    pub(crate) use transactional_blocking;
     pub(crate) use transactional_blocking_with_retry;
 }