+++ sdk/deepbook/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mysten/deepbook
 
+## 0.7.1
+
+### Patch Changes
+
+- 4830361fa4: Updated typescript version
+- Updated dependencies [4830361fa4]
+  - @mysten/sui.js@0.50.1
+
 ## 0.7.0
 
 ### Minor Changes