+++ sdk/suins-toolkit/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @mysten/suins-toolkit
 
+## 0.4.3
+
+### Patch Changes
+
+- 4830361fa4: Updated typescript version
+- Updated dependencies [4830361fa4]
+  - @mysten/sui.js@0.50.1
+
 ## 0.4.2
 
 ### Patch Changes